libnm-util: rename symbol to fix export check
The export check looks for symbols that begin with "nm_" (except
for GObject class stuff) which indicate that the symbol should be
exported.  If the symbols is 'l' (local) as reported by objdump
that probably means we forgot to add it to the .ver file and it
really should be exported.  nm_gvalue_dup() shouldn't be exported
though, so rename it such that it's obviously internal.