settings: fix refcounting when deleting a default wired connection (lp:797868)
When a default wired connection is saved, it gets deleted first since it
has to be re-created using a settings plugin.  But with the settings
rework in 0.9, default wired refcounting changed and the default
wired connection wasn't kept alive in default_wired_try_update()
over the removal/readd.  This caused a use-after-free.