wifi: clarify supplicant disconnect handling behavior
No functional change, but clarify what's supposed to happen when
a disconnect occurs during activation or when the device is activated.

If the connection is an 802.1x or LEAP connection who's password is
NOT_SAVED, then we should request a new password immediately since
the existing one is probably no longer valid (ie, a OTP or token).

Otherwise, start a timer to give the supplicant some time to reconnect.