core: handle N900 USB network interface again
839c7909 made NM ignore this interface based on faulty
information about what the interface did.  It's actually
just a normal network interface that the N900 can use
to talk to the host or whatever.  It's a bit annoying
that for the most part it won't be used and that NM will
keep attempting to connect it with DHCP unless the user
changes the connection to be static IP (N900 defaults to
address 192.168.2.15 and expects the computer to be
192.168.2.14 and requires an ifup in it's Xterm app) but
if you have an N900 you're probably more knowledgable
than most.

http://wiki.maemo.org/N900_USB_networking