core: ignore errors when replacing a default route that already exists
If  the route already exists and the kernel tells us that, we
don't need to do anything.  We certainly don't need to warn
about it in the logs.

There was a typo in the IPv6 default route replace function that
ignored a returned error, and thus we didn't suppress the NLE_EXIST
error like we wanted to.  Do the same for the IPv4 default route
while we're at it.