Revert "wifi: fix entering an endless loop"
This reverts commit 84403ab84a0b5a0ce34594e5203b983da855ed0d.

It appears we want to loop until one of the callback says we're
done, while this patch would exit too early (possibly before we'd
received a reply) since nl_recvmsgs() returns 0 on success.  What
we really want do do here is wait for one of the callbacks we've
registered to tell us that they've successfully handled the message
and gotten a reply.